Resume Analyzer/Checker: Tailoring Your Career Path with AI
GO TO http://127.0.0.1:5000/ For local host. For simplicity's sake, I have opted to not use a server for this task. I could have, but for this project, It would not be necessary. Unlock the power of artificial intelligence in career planning with Resume Analyzer. This Flask application utilizes BERT embeddings to provide a comprehensive analysis of your resume, comparing it to a sample job description. Discover your overall fit and information similarity, gaining valuable insights to enhance your professional journey. Upload your resume now and harness the potential of cutting-edge technology for career optimization!